chore: remove legacy dp registration
