+++ crates/aisix-server/src/telemetry.rs
@@ -16,8 +16,7 @@
 //!   correlation.
 //!
 //! mTLS: the sender presents the same on-disk bundle the heartbeat
-//! worker uses (the bundle persisted by `register::register_and_persist`
-//! at first boot). cp-api derives `env_id` and `dp_id` from the peer
+//! worker uses. cp-api derives `env_id` and `dp_id` from the peer
 //! cert SAN URI, so the request body doesn't carry them — same wire
 //! shape as `/dp/heartbeat`.
